requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
assertpy==1.1
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
exceptiongroup==1.1.1
idna==3.4
iniconfig==2.0.0
jsonschema==4.17.3
packaging==23.1
pluggy==1.2.0
pyrsistent==0.19.3
pytest==7.4.0
requests==2.31.0
tomli==2.0.1
urllib3==2.0.3
